refactor: put icons in assets
