+++ src/assets/icons/index.js
@@ -1,0 +1,7 @@
+export { default as ArrowIcon } from "./ArrowIcon";
+export { default as CalendarIcon } from "./CalendarIcon";
+export { default as ClockIcon } from "./ClockIcon";
+export { default as LinkedinIcon } from "./LinkedInIcon";
+export { default as MapPinIcon } from "./MapPinIcon";
+export { default as TwitterIcon } from "./TwitterIcon";
+export { default as XIcon } from "./XIcon";
